requirements.txt
asgiref==3.6.0
Django==4.2
django-admin-sortable2==2.1.8
django-appconf==1.0.5
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
